refactor: cleanup and streamline triton decoder
- cleanup .so builder logic
- remove nms option
- merge files together
