+++ pennylane/backline/decoders/triton/decoder_frontend.py
@@ -16,27 +16,25 @@
 
 from __future__ import annotations
 
+import shutil
+import tempfile
 from pathlib import Path
 
 import numpy as np
 from numpy.typing import ArrayLike
 
-from .decoder_kernel import _persistent_css_decoder_kernel
-from .triton_shared import build_shared
+from .persistent_kernel import _persistent_css_decoder_kernel
+from .triton_so_builder import build_so
 
 
 def build_css_decoder(
     Hx: ArrayLike,
     Hz: ArrayLike,
     *,
-    bp_variant: str = "sum_product",
     postprocess: str = "osd",
     niter: int = 10,
     prob: float = 0.1,
-    alpha: float = 0.75,
     platform: str = "hip:gfx90a:64",
-    build_dir: str = "decoder_build_dir",
-    library_name: str = "librdma_triton_decoder.so",
     num_warps: int = 1,
     num_stages: int = 1,
     compiler: str = "",
@@ -58,42 +56,32 @@ def build_css_decoder(
         >>> so_path, symbol_name = build_css_decoder(
         ...     Hx,
         ...     Hz,
-        ...     bp_variant="sum_product",
         ...     postprocess="osd",
         ...     niter=10,
         ...     platform="hip:gfx90a:64",
-        ...     build_dir="build",
-        ...     library_name="librdma_triton_decoder.so",
         ... )
 
     Args:
         Hx (ArrayLike): X parity-check matrix.
         Hz (ArrayLike): Z parity-check matrix.
-        bp_variant (str): Belief-propagation variant to run, either
-            ``"sum_product"`` or ``"norm_min_sum"``.
         postprocess (str): Postprocessing step applied after belief propagation. Use
             ``"hard"`` for hard-decision output or ``"osd"`` for ordered-statistics decoding.
         niter (int): Number of decoder iterations.
         prob (float): Uniform prior error probability across qubits.
-        alpha (float): Scaling factor for normalized min-sum decoding.
         platform (str): Triton target string of the form ``"backend:arch:warp_size"``.
             For instance ``"hip:gfx90a:64"`` targets AMD MI200-class GPUs via the
             HIP backend, gfx90a architecture, and warp size 64, while
             ``"cuda:80:32"`` means CUDA backend, SM80 architecture, warp size 32.
-        build_dir (str): Directory used during compilation. The compiled shared library is written
-            here and temporary wrapper files are cleaned up afterwards.
-        library_name (str): Filename of the compiled shared library.
         num_warps (int): Triton kernel launch warp count.
         num_stages (int): Triton pipeline stage count.
         compiler (str): Optional compiler executable override.
         cflags (tuple[str, ...]): Extra compiler flags.
 
     Returns:
-        tuple[Path, str]: Path to the compiled shared library and the exported entrypoint name.
+        tuple[Path, str]: Path to the compiled shared library in a temporary location and the
+            Triton-generated exported entrypoint name. The caller owns the returned file.
     """
-    out = f"{build_dir}/{library_name}"
     _validate_options(
-        bp_variant=bp_variant,
         postprocess=postprocess,
         niter=niter,
         prob=prob,
@@ -101,40 +89,41 @@ def build_css_decoder(
         num_stages=num_stages,
         platform=platform,
     )
-    hx = _normalize_h(Hx)
-    hz = _normalize_h(Hz)
-    return build_shared(
-        _persistent_css_decoder_kernel,
-        signature={
-            "ring_u64_ptr": "*u64",
-            "handoff_u64_ptr": "*u64",
-            "stop_u32_ptr": "*u32",
-            "total": "u64",
-        },
-        constexpr={
-            "Hx": tuple(tuple(int(v) for v in row) for row in hx.tolist()),
-            "Hz": tuple(tuple(int(v) for v in row) for row in hz.tolist()),
-            "BP_VARIANT": bp_variant,
-            "POSTPROCESS": postprocess,
-            "PROB": prob,
-            "NITER": niter,
-            "ALPHA": alpha,
-        },
-        grid=(1, 1, 1),
-        target=platform.strip(),
-        build_dir=str(Path(build_dir).resolve()),
-        out=str(Path(out).resolve()),
-        num_warps=num_warps,
-        num_stages=num_stages,
-        compiler=compiler,
-        cflags=tuple(cflags),
-    )
+    hx = _to_numpy(Hx)
+    hz = _to_numpy(Hz)
+    tmpdir = Path(tempfile.mkdtemp(prefix="pl_triton_decoder_"))
+    out = tmpdir / "librdma_triton_decoder.so"
+    try:
+        return build_so(
+            _persistent_css_decoder_kernel,
+            signature={
+                "ring_u64_ptr": "*u64",
+                "handoff_u64_ptr": "*u64",
+                "stop_u32_ptr": "*u32",
+                "total": "u64",
+            },
+            constexpr={
+                "Hx": tuple(tuple(int(v) for v in row) for row in hx.tolist()),
+                "Hz": tuple(tuple(int(v) for v in row) for row in hz.tolist()),
+                "POSTPROCESS": postprocess,
+                "PROB": prob,
+                "NITER": niter,
+            },
+            grid=(1, 1, 1),
+            target=platform.strip(),
+            out=str(out.resolve()),
+            num_warps=num_warps,
+            num_stages=num_stages,
+            compiler=compiler,
+            cflags=tuple(cflags),
+        )
+    except Exception:
+        shutil.rmtree(tmpdir, ignore_errors=True)
+        raise
 
 
-def _normalize_h(H: ArrayLike) -> np.ndarray:
-    """
-    Ensure matrix is compliant: 2 dimensional of size, filled with 1s and 0s
-    """
+def _to_numpy(H: ArrayLike) -> np.ndarray:
+    """Validate and convert a parity-check matrix to a NumPy array."""
     h = np.asarray(H)
     if h.ndim != 2:
         raise ValueError(f"H must be a 2D array, got shape {h.shape!r}")
@@ -156,7 +145,6 @@ def _normalize_h(H: ArrayLike) -> np.ndarray:
 
 def _validate_options(
     *,
-    bp_variant: str,
     postprocess: str,
     niter: int,
     prob: float,
@@ -165,8 +153,6 @@ def _validate_options(
     platform: str,
 ) -> None:
     """Validate decoder build options."""
-    if bp_variant not in {"sum_product", "norm_min_sum"}:
-        raise ValueError("bp_variant must be 'sum_product' or 'norm_min_sum'")
     if postprocess not in {"hard", "osd"}:
         raise ValueError("postprocess must be 'hard' or 'osd'")
     if niter <= 0:
